lib: loading: implementation of loader
This translates from "syntax" to "semantics", but does not perform any
booking logic.

This is currently very simple, it just resolves date specs to actual
dates (by adding the timezone where missing), but eventually it will be
used to resolve `load` directives, keep track of active options, etc.

We keep the parse and load steps separate to allow writing tooling that
works with the source files, like formatters, linters, etc.