lib: utils: implement `isHeap`, `makeHeap` and `popHeap` functions
The FIFO (LIFO) booking method will need to find the oldest (newest)
positions in the inventory. A naive implementation would just sort the
positions by date, and iterate them in sorted order. But this has
`O(n log n)` complexity in the size of the inventory, which is wasteful
if we only need to reduce a small number (`m < n`) of positions.

Instead, we can build a heap, which has `O(n)` complexity, and pop
elements from it, which has `O(log n)` complexity, resulting in
`O(n + m log n)` complexity, which for small and bounded `m` reduces to
`O(n)`. In the worst case, `m = n` and we would be back to the
performance of the naive implementation.