lib: parsing: fix error handling for syntax errors
When tokenization fails anywhere except on the first token, the error is
not raised immediately, as tokenization happens lazily. Instead, the
error is raised later (as tokens are consumed by the parser). This
failure was not being handled correctly, causing an uncaught exception
printing a stack trace instead of the pretty syntax error message.