lib: add new partitioning functions
This adds the `partitionPoint` function which is a more generic form of
`lowerBound`. In fact, both `lowerBound` and `upperBound` can be
implemented in terms of `partitionPoint`, and are changed here to do so.

Additionally, add `partition`, `partitionLo` and `partitionHi` utilities
to partition the array at the partition point.

Use these new functions in place of calls to `slice` and `lowerBound`
which are harder to reason about.

Note: there are no more usages of `lowerBound` (nor `upperBound`, which
was unused even before) outside of tests. We could remove them, but
maybe they will be useful later. It's also useful to see that their
tests keep passing when reimplemented in terms of `partitionPoint`,
which validates its correctness. So I'm keeping them for now.